requirements.txt
streamlit
requests
streamlit-extras
cryptography
pandas
feedparser
python-dateutil
streamlit-authenticator
pyyaml